[arcane,core+utils] Add specific header for 'MeshVectorMDVariable' and 'MeshMatrixMDVariable'.
Also add specific views for Data setter/getter of `NumMatrix`.
