+++ arcane/src/arcane/core/MeshMDVariableRef.h
@@ -216,278 +216,6 @@ class MeshMDVariableRefT
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
-/*!
- * \brief Class managing a multi-dimensional NumVector type variable on a mesh entity.
- *
- * \warning You need to call the method reshape() before using this kind
- * of variables.
- *
- * For more information, see \ref arcanedoc_core_types_axl_md_variable_use.
- */
-template <typename ItemType, typename DataType, int Size, typename Extents>
-class MeshVectorMDVariableRefT
-: public MeshMDVariableRefBaseT<ItemType, DataType, typename Extents::template AddedFirstLastExtentsType<DynExtent, Size>>
-{
- public:
-
-  using NumVectorType = NumVector<DataType, Size>;
-
- private:
-
-  using BasicType = typename DataTypeTraitsT<DataType>::BasicType;
-  using AddedFirstLastExtentsType = typename Extents::template AddedFirstLastExtentsType<DynExtent, Size>;
-  using AddedFirstExtentsType = typename Extents::template AddedFirstExtentsType<DynExtent>;
-  using BaseClass = MeshMDVariableRefBaseT<ItemType, DataType, AddedFirstLastExtentsType>;
-  static_assert(Extents::rank() >= 0 && Extents::rank() <= 2, "Only Extents of rank 0, 1 or 2 are implemented");
-  static_assert(std::is_same_v<DataType, BasicType>, "DataType should be a basic type (Real, Int32, Int64, ... )");
-
- public:
-
-  using ItemLocalIdType = typename ItemType::LocalIdType;
-  using ReferenceType = DataViewGetterSetter<NumVectorType>;
-  using ConstReferenceType = DataViewGetter<NumVectorType>;
-  using MDSpanType = MDSpan<NumVectorType, AddedFirstExtentsType, RightLayout>;
-  static constexpr int nb_dynamic = Extents::nb_dynamic;
-
- public:
-
-  explicit MeshVectorMDVariableRefT(const VariableBuildInfo& b)
-  : BaseClass(b)
-  {}
-
- public:
-
-  //! \name Operations for variable of dimension MDDim0
-  ///@{
-  //! Accesses the data for reading/writing
-  ReferenceType operator()(ItemLocalIdType id)
-  requires(Extents::rank() == 0)
-  {
-    return ReferenceType(m_vector_mdspan.ptrAt(id.localId()));
-  }
-
-  //! Accesses the data for reading
-  ConstReferenceType operator()(ItemLocalIdType id) const
-  requires(Extents::rank() == 0)
-  {
-    return ConstReferenceType(m_vector_mdspan.ptrAt(id.localId()));
-  }
-  ///@}
-
-  //! \name Operations for variable of dimension MDDim1
-  ///@{
-  //! Accesses the data for reading/writing
-  ReferenceType operator()(ItemLocalIdType id, Int32 i1)
-  requires(Extents::rank() == 1)
-  {
-    return ReferenceType(m_vector_mdspan.ptrAt(id.localId(), i1));
-  }
-
-  //! Accesses the data for reading
-  ConstReferenceType operator()(ItemLocalIdType id, Int32 i1) const
-  requires(Extents::rank() == 1)
-  {
-    return ConstReferenceType(m_vector_mdspan.ptrAt(id.localId(), i1));
-  }
-  ///@}
-
-  //! \name Operations for variable of dimension MDDim2
-  ///@{
-  //! Accesses the data for reading/writing
-  ReferenceType operator()(ItemLocalIdType id, Int32 i1, Int32 i2)
-  requires(Extents::rank() == 2)
-  {
-    return ReferenceType(m_vector_mdspan.ptrAt(id.localId(), i1, i2));
-  }
-
-  //! Accesses the data for reading
-  ConstReferenceType operator()(ItemLocalIdType id, Int32 i1, Int32 i2) const
-  requires(Extents::rank() == 2)
-  {
-    return ConstReferenceType(m_vector_mdspan.ptrAt(id.localId(), i1, i2));
-  }
-  ///@}
-
-  /*!
-   * \brief Changes the data shape.
-   *
-   * The number of elements in \a dims must correspond to the number of dynamic values
-   * in \a Extents.
-   */
-  void reshape(std::array<Int32, Extents::nb_dynamic> dims)
-  {
-    std::array<Int32, nb_dynamic + 1> full_dims;
-    // We add 'Size' to the end of the dimensions.
-    for (int i = 0; i < nb_dynamic; ++i)
-      full_dims[i] = dims[i];
-    full_dims[nb_dynamic] = Size;
-    ArrayShape shape(full_dims);
-    this->m_underlying_var.resizeAndReshape(shape);
-  }
-
- protected:
-
-  void updateFromInternal() override
-  {
-    BaseClass::updateFromInternal();
-    // Positions the value of m_vector_mdspan.
-    // It will have the same dimensions as m_mdspan except that we
-    // remove the last dimension and change the type
-    // from 'DataType' to 'NumVector<DataType,Size>'.
-    DataType* v = this->m_mdspan.to1DSpan().data();
-    NumVectorType* nv = reinterpret_cast<NumVectorType*>(v);
-    m_vector_mdspan = MDSpanType(nv, this->m_mdspan.extents().dynamicExtents());
-  }
-
- private:
-
-  MDSpanType m_vector_mdspan;
-};
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-/*!
- * \brief Class managing a multi-dimensional NumMatrix type variable on a
- * mesh entity.
- *
- * The dimension of the matrix is fixed and is given by (Row,Column).
- *
- * \warning You need to call the method reshape() before using this kind
- * of variables.
- *
- * For more information, see \ref arcanedoc_core_types_axl_md_variable_use.
- */
-template <typename ItemType, typename DataType_, int Row, int Column, typename Extents>
-class MeshMatrixMDVariableRefT
-: public MeshMDVariableRefBaseT<ItemType, DataType_, typename Extents::template AddedFirstLastLastExtentsType<DynExtent, Row, Column>>
-{
- public:
-
-  using DataType = DataType_;
-  using NumMatrixType = NumMatrix<DataType, Row, Column>;
-
- private:
-
-  using BasicType = typename DataTypeTraitsT<DataType>::BasicType;
-  using AddedFirstLastLastExtentsType = typename Extents::template AddedFirstLastLastExtentsType<DynExtent, Row, Column>;
-  using AddedFirstExtentsType = typename Extents::template AddedFirstExtentsType<DynExtent>;
-  using BaseClass = MeshMDVariableRefBaseT<ItemType, DataType, AddedFirstLastLastExtentsType>;
-  static_assert(Extents::rank() >= 0 && Extents::rank() <= 1, "Only Extents of rank 0 or 1 are implemented");
-  static_assert(std::is_same_v<DataType, BasicType>, "DataType should be a basic type (Real, Int32, Int64, ... )");
-
- public:
-
-  using ItemLocalIdType = typename ItemType::LocalIdType;
-  using ReferenceType = DataViewGetterSetter<NumMatrixType>;
-  using ConstReferenceType = DataViewGetter<NumMatrixType>;
-  using MDSpanType = MDSpan<NumMatrixType, AddedFirstExtentsType, RightLayout>;
-  static constexpr int nb_dynamic = Extents::nb_dynamic;
-
- public:
-
-  explicit MeshMatrixMDVariableRefT(const VariableBuildInfo& b)
-  : BaseClass(b)
-  {}
-
- public:
-
-  //! \name Operations for variable of dimension MDDim0
-  ///@{
-  //! Mutable view of the matrix for item \a id
-  ReferenceType operator()(ItemLocalIdType id) requires(Extents::rank()==0)
-  {
-    return ReferenceType(m_matrix_mdspan.ptrAt(id.localId()));
-  }
-
-  //! Read-only view of the matrix for item \a id
-  ConstReferenceType operator()(ItemLocalIdType id) const requires(Extents::rank()==0)
-  {
-    return ReferenceType(m_matrix_mdspan.ptrAt(id.localId()));
-  }
-
-  //! Mutable view of the element (i,j) of the matrix for item \a id
-  DataType& operator()(ItemLocalIdType id, Int32 i, Int32 j) requires(Extents::rank() == 0)
-  {
-    return m_matrix_mdspan(id.localId())(i, j);
-  }
-
-  //! Read-only view of the element (i,j) of the matrix for item \a id
-  DataType operator()(ItemLocalIdType id, Int32 i, Int32 j) const requires(Extents::rank() == 0)
-  {
-    return m_matrix_mdspan(id.localId())(i, j);
-  }
-  ///@}
-
-  //! \name Operations for variable of dimension MDDim1
-  ///@{
-  //! Mutable view of the matrix of index \a index for item \a id
-  ReferenceType operator()(ItemLocalIdType id, Int32 index)
-  requires(Extents::rank() == 1)
-  {
-    return ReferenceType(m_matrix_mdspan.ptrAt(id.localId(), index));
-  }
-
-  //! Read-only view of the matrix of index \a index for item \a id
-  ConstReferenceType operator()(ItemLocalIdType id, Int32 index) const
-  requires(Extents::rank() == 1)
-  {
-    return ReferenceType(m_matrix_mdspan.ptrAt(id.localId(), index));
-  }
-
-  //! Mutable view of the element (i,j) of the matrix for item \a id and index \a index
-  DataType& operator()(ItemLocalIdType id, Int32 index, Int32 i, Int32 j)
-  requires(Extents::rank() == 1)
-  {
-    return m_matrix_mdspan(id.localId(), index)(i, j);
-  }
-
-  //! Read-only view of the element (i,j) of the matrix for item \a id and index \a index
-  DataType operator()(ItemLocalIdType id, Int32 index, Int32 i, Int32 j) const
-  requires(Extents::rank() == 1)
-  {
-    return m_matrix_mdspan(id.localId(), index)(i, j);
-  }
-  ///@}
-
-  /*!
-   * \brief Changes the data shape.
-   *
-   * The number of elements in \a dims must correspond to the number of dynamic values
-   * in \a Extents.
-   */
-  void reshape(std::array<Int32, Extents::nb_dynamic> dims)
-  {
-    std::array<Int32, nb_dynamic + 2> full_dims;
-    // We add 'Row' and 'Column' to the end of the dimensions.
-    for (int i = 0; i < nb_dynamic; ++i)
-      full_dims[i] = dims[i];
-    full_dims[nb_dynamic] = Row;
-    full_dims[nb_dynamic + 1] = Column;
-    ArrayShape shape(full_dims);
-    this->m_underlying_var.resizeAndReshape(shape);
-  }
-
- protected:
-
-  void updateFromInternal() override
-  {
-    BaseClass::updateFromInternal();
-    // Positions the value of m_vector_mdspan.
-    // It will have the same dimensions as m_mdspan except that we
-    // remove the last dimension and change the type
-    // from 'DataType' to 'NumMatrix<DataType,Row,Column>'.
-    DataType* v = this->m_mdspan.to1DSpan().data();
-    NumMatrixType* nv = reinterpret_cast<NumMatrixType*>(v);
-    m_matrix_mdspan = MDSpanType(nv, this->m_mdspan.extents().dynamicExtents());
-  }
-
- private:
-
-  MDSpanType m_matrix_mdspan;
-};
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
 
 } // namespace Arcane
