+++ arcane/src/arcane/utils/NumMatrix.cc
@@ -1,0 +1,15 @@
+﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
+//-----------------------------------------------------------------------------
+// Copyright 2000-2026 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// See the top-level COPYRIGHT file for details.
+// SPDX-License-Identifier: Apache-2.0
+//-----------------------------------------------------------------------------
+/*---------------------------------------------------------------------------*/
+/* NumMatrix.cc                                                (C) 2000-2026 */
+/*                                                                           */
+/* Fixed-size mathematical matrix of numeric types.                          */
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+#include "arcane/utils/NumMatrix.h"
+#include "arcane/utils/NumMatrixDataView.h"
